requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
google-api-python-client==1.7.11
google-auth==1.10.1
google-auth-httplib2==0.0.3
httplib2==0.16.0
idna==2.8
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
pyTelegramBotAPI==3.6.7
PyYAML==5.3
requests==2.22.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.7
